toolkit/devtools/server/tests/mochitest/test_inspector-remove.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
Components.utils.import("resource://gre/modules/devtools/Loader.jsm");
const {Promise: promise} = Components.utils.import("resource://gre/modules/Promise.jsm", {});

const inspector = devtools.require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gClient = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = devtools.require("devtools/server/actors/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addTest(function testRemoveSubtree() {
  let originalOwnershipSize = 0;
  let longlist = null;
  let longlistID = null;

  let nextSibling = gInspectee.querySelector("#longlist").nextSibling;
  // Duplicate the walker logic to skip blank nodes...
  while (nextSibling && nextSibling.nodeType === Components.interfaces.nsIDOMNode.TEXT_NODE && !/[^\s]/.exec(nextSibling.nodeValue)) {
    nextSibling = nextSibling.nextSibling;
  }

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#longlist").then(listFront => {
    longlist = listFront;
    longlistID = longlist.actorID;
  }).then(() => {
    return gWalker.children(longlist);
  }).then((items)=> {
    originalOwnershipSize = assertOwnership();
    ok(originalOwnershipSize > 26, "Should have at least 26 items in our ownership tree");
    return gWalker.removeNode(longlist);
  }).then(nextSiblingFront => {
    is(nextSiblingFront.rawNode(), nextSibling, "Should have returned the next sibling.");
    return waitForMutation(gWalker, isChildList);
  }).then(() => {
    // Our ownership size should now be 26 fewer (we forgot about #longlist + 26 children, but learned about #longlist's next sibling)
    let newOwnershipSize = assertOwnership();
    is(newOwnershipSize, originalOwnershipSize - 26, "Ownership tree should have dropped by 27 nodes");
    // Now verify that some nodes have gone away
    return checkMissing(gClient, longlistID);
  }).then(runNextTest));
});

addTest(function cleanup() {
  delete gWalker;
  delete gClient;
  runNextTest();
});


  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
